+++ __tests__/example.test.tsx
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
 
-import Home from "../app/(route)/(admin)/page";
+import Home from "../app/(route)/(admin)/zuzuclubadmin/page";
 
 describe("Home", () => {
   it("renders single div tag", () => {
