+++ app/(route)/(admin)/zuzuclubadmin/[id]/(hooks)/useAlimTable.tsx
@@ -15,9 +15,9 @@ import {
 } from "@tanstack/react-table";
 import type { Table } from "@tanstack/react-table";
 
-import { AlimTHeaderColumn } from "../../../../components/root/AlimColumn";
-import { useGetAcceptance } from "../../../../hooks/query";
-import { AcceptanceSchema } from "../../../../actions/get-acceptance";
+import { AlimTHeaderColumn } from "../../../../../components/admin/AlimColumn";
+import { useGetAcceptance } from "../../../../../hooks/query";
+import { AcceptanceSchema } from "../../../../../actions/get-acceptance";
 
 interface AlimTableContextType {
   alimTable: Table<AcceptanceSchema["alarmTalkResponses"][0]>;
