+++ app/(route)/(admin)/zuzuclubadmin/[id]/page.tsx
@@ -1,7 +1,7 @@
-import { Header } from "../../../ui";
-import { Alim } from "../../../components/root/Alim";
-import ParentsRequest from "../../../components/root/ParentsRequest";
-import { Teacher } from "../../../components/root/Teacher";
+import { Header } from "../../../../ui";
+import { Alim } from "../../../../components/admin/Alim";
+import ParentsRequest from "../../../../components/admin/ParentsRequest";
+import { Teacher } from "../../../../components/admin/Teacher";
 
 export default function MatchingDetailHome({
   params,
