+++ app/(route)/(admin)/zuzuclubadmin/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginPage from "../../../components/auth/login";
+import LoginPage from "../../../../components/auth/login";
 
 export default function Home() {
   return <LoginPage />;
