+++ app/(route)/page.tsx
@@ -1,0 +1,3 @@
+export default function Home() {
+  return <div className="w-full">y-edu임</div>;
+}
