+++ app/components/admin/AlimHeader.tsx
@@ -10,7 +10,7 @@ import { useGetAcceptance } from "../../hooks/query";
 import { usePostMatchingAcceptance } from "../../hooks/mutation";
 import { useModal } from "../../hooks/custom";
 import { Modal } from "../../ui";
-import { useAlimTableContext } from "../../(route)/(admin)/[id]/(hooks)/useAlimTable";
+import { useAlimTableContext } from "../../(route)/(admin)/zuzuclubadmin/[id]/(hooks)/useAlimTable";
 
 interface AlimHeaderProps {
   matchingId: string;
